refactor(api): git commits 라우트를 전용 모듈로 추출
5개 라우트(commits 4 + draft-commits 1)를 api/git/commits.py로 이동.
핸들러 본문·경로·함수명 무변경 — sed 로 원문 구간을 그대로 잘라 옮겼다.
라우트 테이블 characterization 테스트 통과 — HTTP 표면 63건 동일, 가림 쌍 0건.

_get_llm_access_for_git 헬퍼도 함께 이동했다(사용처가 draft-commits
핸들러 한 곳뿐이라 _shared.py 승격 대상이 아니다). 이에 따라 _legacy.py 의
LLM 관련 import 4종과 commit/draft 모델 6종이 미사용이 되어 제거했다.

__init__.py 의 비-router 재노출 출처를 ._legacy → .commits 로 전환했다.
Task 3~6 은 router 만 다뤘으므로 이 경로는 처음 검증된다 —
generate_draft_commits.__module__ == 'api.git.commits' 로 런타임 확인.

테스트 문자열 패치 타깃 2곳을 api.git._legacy.get_git_service_for_project
→ api.git.commits.get_git_service_for_project 로 갱신했다. 이로써
_legacy 를 가리키는 패치 지점은 0건이 됐다.
